Update edit UI to use differnet canvas. Allow for Creation and deletion. Editing records WIP
