+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "basic-gui"
+name = "copycat"
 version = "0.1.0"
 edition = "2021"
 
@@ -10,5 +10,4 @@ clipboard = "0.5.0"
 device_query = "1.1.2"
 eframe = "0.21.3"
 egui = "0.21.0"
-sccache = "0.3.3"
 x11-clipboard = "0.7.0"
